requirements.txt
django==2.2.13
django_cors_headers
graphene_django==2.8.2
django_random_id_model
pyjwt==1.7.0
requests
graphene_file_upload
django-cleanup
django-timezone-field
django-environ